Features/exactsolver (#155)
Added exact solver:

For selected models that have only relatively few states, this allows you to compute the optimal policy and costs. Examples python/scripts/dp_exact_solver.py and in tests/other_unit_tests/t_exact_solver_tests.cpp. 

Note that it is no longer needed (like it was in old DP version) to give allstates. But is is needed to implement:

	std::vector<std::tuple<Event, double>> EventProbabilities() const;
		
This may also be state-dependent, i.e. with parameter state. Other than that, MDP must implement GetFeatures, as states with equal features are considered to be equal by the policy iteration algorithm. 

lost_sales model satisfies requirements for this.

---------

Co-authored-by: willem van jaarsveld <wvanjaarsveld@hotmail.com>